make CAPG gaussian the default
